+++ lib/SearchAndSort/SearchAndSort.js
@@ -17,6 +17,7 @@ import Button from '@folio/stripes-components/lib/Button';
 import MultiColumnList from '@folio/stripes-components/lib/MultiColumnList';
 import Layer from '@folio/stripes-components/lib/Layer';
 import SRStatus from '@folio/stripes-components/lib/SRStatus';
+import transitionToParams from '@folio/stripes-components/util/transitionToParams';
 import UserForm from '../../UserForm'; // XXX relocate this
 import ViewUser from '../../ViewUser'; // XXX relocate this
 import contactTypes from '../../data/contactTypes';
@@ -111,6 +112,7 @@ class SearchAndSort extends React.Component {
       showNotesPane: false,
     };
 
+    this.transitionToParams = transitionToParams.bind(this);
     this.commonChangeFilter = commonChangeFilter.bind(this);
     this.connectedViewUser = props.stripes.connect(ViewUser);
     this.SRStatus = null;
@@ -131,6 +133,23 @@ class SearchAndSort extends React.Component {
       const resultAmount = nextProps.parentResources.users.other.totalRecords;
       this.SRStatus.sendMessage(`Search returned ${resultAmount} result${resultAmount !== 1 ? 's' : ''}`);
     }
+  }
+
+  onSort = (e, meta) => {
+    const newOrder = meta.alias;
+    const oldOrder = this.state.sortOrder || '';
+
+    const orders = oldOrder ? oldOrder.split(',') : [];
+    if (orders[0] && newOrder === orders[0].replace(/^-/, '')) {
+      orders[0] = `-${orders[0]}`.replace(/^--/, '');
+    } else {
+      orders.unshift(newOrder);
+    }
+
+    const sortOrder = orders.slice(0, 2).join(',');
+    this.log('action', `sorted by ${sortOrder}`);
+    this.setState({ sortOrder });
+    this.transitionToParams({ sort: sortOrder });
   }
 
   onChangeFilter = (e) => {
